refactor: aplicação de padrões profissionais e precisão monetária
- Migração de valores monetários de Float para Long (centavos) para evitar erros de precisão decimal.
- Implementação de Kotlin Extension Functions para formatação de moeda, data e utilitários de UI.
- Refatoração do FirebaseHelper para tratar CancellationException e suporte a inversão de dependência.
- Remoção de classes utilitárias estáticas (GetMask) em favor de extensões nativas.
- Atualização do MoneyTextWatcher para manipulação de centavos.

+++ app/src/main/java/com/example/megaburguer/data/model/Menu.kt
@@ -7,7 +7,7 @@ import kotlinx.parcelize.Parcelize
 data class Menu(
     val id: String = "",
     val nameItem: String = "",
-    val price: Float = 0f,
+    val price: Long = 0L,
     val category: String = ""
 
 ): Parcelable
